+++ src/app/project/project-documents/project-documents.component.ts
@@ -71,6 +71,7 @@ export class ProjectDocumentsComponent implements OnInit, OnDestroy {
   public canUnpublish;
   public pathAPI: string;
   public tableParams: TableParamsObject = new TableParamsObject();
+  public focusRef: HTMLElement | null;
 
   constructor(
     private _changeDetectionRef: ChangeDetectorRef,
@@ -141,11 +142,27 @@ export class ProjectDocumentsComponent implements OnInit, OnDestroy {
         }
       });
 
-    this.ngxSmartModalService.getModal('confirmation-modal').onAnyCloseEventFinished
+    const confirmationModal = this.ngxSmartModalService.getModal('confirmation-modal');
+    confirmationModal.onOpen
       .takeUntil(this.ngUnsubscribe)
-      .subscribe((modal) => {
-      const data = this.ngxSmartModalService.getModalData('confirmation-modal');
-      this.documentActions(data);
+      .subscribe(() => {
+        // Save the active element so we can return focus.
+        this.focusRef = document.activeElement as HTMLElement | null;
+      })
+
+    confirmationModal.onAnyCloseEvent
+      .takeUntil(this.ngUnsubscribe)
+      .subscribe(() => {
+        // Return focus to the previously-active element.
+        if (this.focusRef) this.focusRef.focus();
+      })
+
+    confirmationModal.onAnyCloseEventFinished
+      // Perform document actions on close.
+      .takeUntil(this.ngUnsubscribe)
+      .subscribe(() => {
+        const data = this.ngxSmartModalService.getModalData('confirmation-modal');
+        this.documentActions(data);
       });
 
     this.pageBreadcrumbs = [{ pageTitle: this.currentProject.name, routerLink: [ '/p', this.currentProject._id ]}];
@@ -292,6 +309,7 @@ export class ProjectDocumentsComponent implements OnInit, OnDestroy {
         Promise.all(promises).then(() => {
           this.openSnackBar('Download initiated.', 'Close');
         });
+        break;
       case 'publish':
         this.onPublishDocument();
         break;