+++ src/app/project/project-email-subscribe/email-subscribe-table-rows/email-subscribe-table-rows.component.ts
@@ -68,4 +68,13 @@ export class EmailSubscribeTableRowsComponent implements OnInit, TableComponent 
         });
   }
 
+  /**
+   * Terminate subscriptions when component is unmounted.
+   *
+   * @return {void}
+   */
+  ngOnDestroy() {
+    this.ngUnsubscribe.next();
+    this.ngUnsubscribe.complete();
+  }
 }